LUN-189: Split CustomDisciplinMenuItem into two componets
